Sync rides with the database

+++ MtbMate/Database/Create/Tables/Jump.sql
@@ -1,10 +1,10 @@
 ﻿CREATE TABLE [Jump] (
     [JumpId] int NOT NULL IDENTITY(1,1) PRIMARY KEY,
 
-	[RideId] int NOT NULL CONSTRAINT [FK_Jump_Ride] REFERENCES [Ride],
+    [RideId] int NOT NULL CONSTRAINT [FK_Jump_Ride] REFERENCES [Ride],
 
-	[Number] int NOT NULL,
-	[Time] datetime NOT NULL,
-	[Airtime] decimal NOT NULL,
-	[LandingGForce] decimal NOT NULL,
+    [Number] int NOT NULL,
+    [Time] datetime NOT NULL,
+    [Airtime] decimal(5,3) NOT NULL,
+    [LandingGForce] decimal(5,3) NOT NULL,
 )
